module: use uintptr_t as type for kernel symbols withuot type info
When a kernel symbol is known to exist (found in kallsyms) but no type
information can be found, a compilation error is reported.  That makes
it impossible to use casting to resolve missing type information for
symbols whose address is clearly known.  So, assign those symbol the
uintptr_t type.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>