usdt: remove relocations rather than changing their type to R_*_NONE
The original DTrace code used to depend on STV_ELIMINATE and XXX
to have the linker clean up USDT probe related relocations.  This
does not work on Linux, so DTrace on Linux changed the type of those
relocations to be R_*_NONE since those relocations get ignored.

Binutils no longer allows that for PIC objects when a locally referenced
symbol is involved.

DTrace will now remove those relocations since they are no longer needed
anyway.

A test is included to ensure that PIC objects can be linked successfully
when USDT probes are used.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>