uprobes: name uprobes based on offset in inode rather than address
When PIC code is used, USDT probes (using uprobes) can appear at different
addresses for distinct instances of the same ELF object because the base
address is not fixed.  As a result, uprobes were being created using names
that were not stable.  Since (dev, inode, offset) is a unique triplet for
any uprobe, use that instead of (dev, inode, address).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>