build: support kernel builds with or without object directory
It should be possible to build DTrace against kernel headers provided in
a variety of ways.  They can be provided as installed kernel header
packages with combined or separated generated and standard headers, and
linked from /lib/modules/<kernelv-version> or simply available as
directory hierarchies elsewhere on the system.

It should also be possible to build DTrace against kernel headers that
are found in a locally built kernel, whether it is installed or not.

This patch provides a cleanup for the build system to accomplish this
with minimal configuration.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>