EINTR: make safe under signal hits
This commit audits DTrace (especially libproc and dt_proc) for syscalls and
library calls that may return EINTR but are not suitably guarded, and guards
them, adjusting any that use timeouts so that the timeout ignores the
interruption.  (The plan was also to have any of them that needed to
actually *use* this EINTR to trigger an early return to do that as well, but
it turns out that none do, not even the ones in libproc.)

This matters for libproc and dt_proc in particular now because the monitor
thread is routinely hit by signals and so has to expect EINTR returns from
appropriate syscalls (and consider what to do if that happens, because it
means a proxy request is waiting and the DTrace main thread is blocked on
it).

This does not touch the core polling loop in dt_consume, since that loop is
*expected* to return -EINTR, triggering an -EINTR return from dtrace_go()
(see cmd/dtrace.c's call to that function).  This may be a bug, since it
means that repeatedly changing the window size is enough to cause DTrace to
ignore the switchrate -- but the caller might well *want* an -EINTR return
so it can handle whatever signal just hit.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>