Add support for cleanpath() subroutine
Common simplifications include:
- replace all "//" with "/"
- replace all "/./" with "/"
- replace all "/foo/.." with "/"
but precise semantics can get tricky in certain cases.

Notice that the results differ in some cases from what was
done in legacy DTrace:

    input string         current patch     legacy DTrace

    "/."                 "/"               "/."
    "a/."                "a"               "a/."
    "./a"                "a"               "./a"

    "../.."              "../.."           "."
    "a/.."               "."               "a"
    "a/../.."            ".."              "a"

    "a/../b"             "b"               "/b"

In the first set, legacy DTrace has unnecessary trailing "." or "/."
or unnecessary leading "./".  They would seem to be wrong.

In the next cases, the legacy results are clearly wrong, not
handling ".." correctly.

In the final case, a relative path is converted into an absolute
path, which is also clearly incorrect.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>