rawtp: retrieve probe arguments from CTF data
We were determining the number of arguments using a trial-and-error
approach with minimal BPF programs.  If there is CTF type data for
the raw tracepoints (in the form of a function prototype), we use
that because it also gives us the types of the arguments.

If no such CTF data is available, we fall back to the trial-and-error
approach.

The included test will fail if the fallback is used.  There is no way
to force the fallback to get triggered explicitly.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>