spec: support building from RPMs again
We need a way to override the /source and /build suffixes which
/lib/modules/$(uname -r) needs after it to find the kernel
headers, because devel rpms don't use either (things are just
directly in /usr/src/kernels/$(kernel)/, with no suffix).

Introduce KERNELSRCNAME and KERNELBLDNAME which the specfile can
set to nothing to make the RPMs build again.

(As a benefit, this might mean we can be made to build fine
without a source/ link at all, which current upstream kernels
are now doing.  Not tested that yet, though.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>