bpf, perf: helper functions should return -errno on failure
The bpf() and perf_event_open() helper functions were returning -1 on
failure and callers were supposed to check errno.  Since the only valid
values are >= 0, it would be convenient to avoid the burden on callers
to check for a -1 return value and to then check errno (which can also be
dangerous if there are intermediary functions that could possibly cause
the errno value to change)..

The functions are now renamed dt_bpf() and dt_perf_event_open() to avoid
confusion with possible C library system call wrappers and they now
return -errno on failure so callers can just check the return value
rather than errno.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>