test: XFAIL lockmem test if rawtp-arg-type uses CTF
To get rawtp arg types, we rely on CTF.  If CTF info is not available,
we fall back on a "trial and error" method simply for the number of args.
But we signal to the user if the trial-and-error method is failing due
to lockmem limits.

This test checked for that failure mode.  The test was skipped if
lockmem limits are not observed on the test system.

Skip the test also if CTF info *IS* available.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>