test: Update -xmodpath test
We test -xmodpath by checking that pointing modpath to:
- an empty directory produces "bad" results
- an appropriate symlink reproduces good (default) results

With the addition of a BTF-to-CTF convertor, however, DTrace basically
functions even in the absence of CTF info.

One signature of the convertor, though, is that rawtp-provider args[]
type information will be (void*) regardless of what the actual type is.
Use this signature to verify that -xmodpath is working.

One complication is that a trial-and-error method of getting rawtp args[]
info is used if there is no HAVE_LIBCTF.  In this case, rawtp args[]
types will be uint64_t, even when no -xmodpath is specified.  Skip the
test in this case.

To review:

- If no HAVE_LIBCTF, rawtp args[] type will always be uint64_t.
  Skip the test in this case.

- If HAVE_LIBCTF but -xmodpath points at an empty directory, many
  DTrace features should still work, but rawtp args[] types will
  all be (void*).  Check that -xmodpath=empty produces (void*).

- Under "normal" circumstances, many rawtp args[] types will be
  neither uint64_t nor (void*).  Check such results in the default
  case and when -xmodpath points at a valid CTF archive.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>