print: Fix truncation
The printsize option can be used to limit how much data is reported
by print().  A few corrections are needed.

1)  In dt_print_visit(), we check if offset/NBBY > dv_size.  E.g.,
if a 40-byte limit is set, then an 8-byte record at offset 40 will
be reported.  Change to offset/NBBY >= dv_size.  We do not worry
about the case where the print size ends in the middle of a record.

2)  dt_print_type() sets the size and then calls dt_print_visit().
But dt_print_visit() can also call itself recursively, and in this
case dv_size is not set.  The dv_size values at deeper depths are
garbage.  Have dt_print_visit() set dva2.dt_size.

The existing test for -xprintsize is corrected and a new test,
focused on truncation at deeper depths, is added.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>