Assigning 0 to any dynamic variable should delete it
Assigning 0 to a scalar dynamic variable deletes it from storage, but
due to type compatibility rules, this could not be used to delete dynamic
variables of struct or union type.

This does not actually change the D language as far as it is documented
(even though the implementation failed to do what was documented):

"[...], assigning an associative array element to zero causes DTrace to
 deallocate the underlying storage."
"Also, as with associative array elements, assigning zero to a
 thread-local variable causes DTrace to deallocate the underlying
 storage."

Add a special case for assigning a literal 0 constant value to any
dynamic variable to signify deleting the variable from storage.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>