Fix drop counter handling for local-only updates
When at the end of a work loop, the drop counters are updated for local
updates only (i.e. when tracing is not stopped yet), an invalid drop
count could be reported if a drop occured in the BEGIN probe.  The cause
is that the code assumed that both copies of the status data contained
valid data.  But when a drop occurs in the BEGIN probe, only the first
copy will have non-zero values.  Since local-only updates assume that
the two copies are identical except for a possible change due to the
local update, and any earlier status update would already have reported
any differences, it is safe to just copy all data prior to performing
the local updates.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>