cg: integrate dvar lookup into dt_cg_prep_dvar()
The loading and storing of dynamic variables was implemented using a
function dt_cg_prep_dvar_args() to prepare the arguments for a function
call, followed by performing the actual function call in the respective
callers.  Those calls are identical, so it makes sense to consolidate
argument setup and call into a single function.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>