Support aggregations of non-void actions
Some actions -- such as stack() and func() -- have non-void actions.
D allows one to use such actions as aggregation keys.  Add such support.

We allow these actions to be keys to aggregations, but not
to associative arrays in order to conform to legacy behavior.
However, lifting this limitation should be easy.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>