Set the lockmem limit to "unlimited" by default
Commit d5a2077d ("Locked-memory limit") described the importance of
setting the locked-memory limit appropriately for DTrace, but it put
that burden on users, admittedly providing several mechanisms to do so.

Driven by commit 5a12c51c ("options: ensure lockmem is set before
retrieving probe info"), commit 7e159efa ("Forbid setting lockmem value
with a pragma") eliminated one such mechanism.

That pragma mechanism, however, had proven to be a popular way of
constructing D scripts that could run both under legacy DTrace, which
does not recognize the lockmem option, and DTrace on Linux, which
essentially requires the option.  That is, the mechanism allowed a
script to set the pragma conditionally based on DTrace version.

Further, it is difficult to size the amount of locked memory needed,
and so users end up just routinely setting the limit to "unlimited."

While there are good reasons not to change resource limits quietly, the
usability issues around lockmem have become nuisances.

Therefore, continue to support the lockmem option, but change the default
behavior to set the lockmem limit to "unlimited."

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>