test: Fix use of syscall::execve:entry args[1][?]
Commit 82332371 ("proc: use a rawtp for the proc:::exit probe") included
some test changes.  Specifically, it sought to use syscall::execve:entry
probe arguments args[1][0] and args[1][1] to recognize "sleep 10000".
The patch recognized that the argv pointers in question were in user
space, requiring copyinstr() to access the strings.

But it's trickier than that.  The args[1][?] require two dereferencings,
both in user space.  So a copyin() is required to access args[1] and
then copyinstr() to access the args[1][?].

Fix the tests to use two layers of copyin*() to double dereference the
args[1][?] strings.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>