dtprobed: fill out the DOF stash
This commit arranges to create DOF stash entries corresponding to the probes
received by dtprobed when a DTRACEHIOC_ADDDOF is received from one of its
DOF-containing CUSE clients, and to remove them again when a
DTRACHIOC_REMOVE is received: at startup and at intervals (every 128
requests, an arbitrary value), the stash is scanned to remove entries
relating to processes that died before they had a chance to send a
DTRACEHIOC_REMOVE.

The stash is located under /run/dtrace by default, or under the directory
given by the new -s option to dtprobed.  This is mostly going to be used by
in-tree testing.  DTrace will gain a new dofstashpath option for the same
reason.  There is no need to document these: like dtprobed's -n option and
the DTRACE_DOF_INIT_DEVNAME environment variable, these are really only
testsuite internal implementation details.

Nothing is yet done with the stash: dtprobed still creates probes itself.

DOF-containing processes doing fork and exec have no effect on the stash as
yet: this will be fixed in a later commit.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>