dt_pid, dtprobed: move uprobe creation to dtrace
Now the DOF stash is being created and populated, we can move all probe
creation to DTrace. This solves multiple problems:

 - removal of probes is difficult: dtprobed knows when all processes that
   registered DOF have unregistered it again, but that doesn't mean it's
   allowed to remove any of the corresponding uprobes. You can't remove a
   uprobe if it has any BPF associated with it, and the thing that does that
   is dtrace itself, which might well remove the BPF long after the
   processes that contained the corresponding uprobes are all dead: so
   dtrace is the only thing that can safely remove uprobes.

 - insertion of uprobes is limiting: passing the DTrace-side probe name down
   via probe arguments was clever, but alas probe arguments have much
   shorter maximum lengths than DTrace probes, so some valid DTrace probes
   cannot be registered as uprobes and end up with truncated names

 - it doesn't scale: the new probe registration/removal grinds to a halt
   after a few tens of thousands of uprobes are registered, and doing it
   here means that every probe that might potentially get used is created,
   even though any given DTrace invocation is only likely to use a subset
   of them.

The solution to this is to remove all the uprobe registration code from
dtprobed (goodbye, libcommon/uprobes.c) and migrate it all into
libdtrace/dt_prov_uprobe.c.  This is convenient because it's where PID probe
registration already happens, and it turns out we can reuse almost all the
code.

As for figuring out which probes to create, instead of pre-scanning the
/sys/kernel/debug/tracing/uprobe_events list and then looking for probes
that relate to the current process, we can just scan the /run/dtrace/probes
subdirectory in the DOF stash that relates to the specific PID we are
interested in (using globbing to identify the probes, taking advantage of
the fact that probespec globs follow the same rules as filesystem globs),
pull in the DOF, and generate probes from it.  Much simpler and much less
work.  (We do need to allow for processes dying as DTrace is processing
stash entries relating to them.)

The pre-parsing has one tiny caveat, which we can't even hit until we start
to reparse probes at runtime rather than just at DTrace startup. All the DOF
we need is already parsed by dtprobed, but if dtprobed was upgraded since
this instance of DTrace started, and the new dtprobed has a newer struct
dof_parsed, we want to ignore any such (too-new) parsed DOF and not register
the corresponding probes.  Earlier commits in this series introduced a new
DOF_PARSED_VERSION #define which is bumped whenever struct dof_parsed
changes and gets stuck at the start of all parsed DOF in the DOF stash: we
just need to compare that to what's baked into this copy of DTrace, and
avoid using any parsed DOF for which those are different.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>