dtprobed: a few test infrastructure improvements
The testsuite needs a few adjustments to cater for the DOF stash.  Firstly,
just as the testsuite instance of dtprobed uses a distinct device node, it
should also use a distinct DOF stash: putting it under the test tmpdir means
that it's even using the same filesystem type (tmpfs) as the DOF stash
usually does, without colliding with anything the systemwide instance may be
doing.

Also make the cleanup interval at which dtprobed cleans up dead internal
userdata structures and scans for kill -9'd processes in the DOF stash
configurable via the already-set _DTRACE_TESTING env var, and reduce it from
its default of 128 to 5, so that the cleanup code gets a good workout.

Also test the reparsing code (which by default would never kick in when the
testsuite ran at all, because the whole testsuite run is done with a single
daemon instance, so daemon upgrades are never relevant) by arranging to
kick off a forced reparse whenever the daemon is hit with a SIGUSR2.  (A
test added in a future commit will use this.)

Finally, avoid prefixing log messages with "dtprobed DEBUG $timestamp: "
when not testing: it's formatted like that to force inclusion in the
runtest.log as a trace message, but it's emitted even for non-debugging
messages so is more than a bit confusing.  Make it a testsuite-only
thing.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>