test: usdt: multitrace: wipe out parsed commits properly
One of the things test/unittest/usdt/tst.multitrace.sh is trying to do is
verify that dtprobed's reparsing of wrong-version parsed DOF works right.
It does this by overwriting the parsed DOF with junk (which corresponds to a
definitely-wrong version) and letting it reparse it.

Unfortunately after the revamp to put the parsed DOF in multiple files,
the test was never adjusted, so it's failing to wipe out the parsed
commits and this never got properly tested since then.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>