consume: ustack: handle errors from dt_Pobjname() better
dt_Pobjname(), which returns the name of the ELF object corresponding to a
specific address, can fail, for instance if the process it's being asked
about is dead.  One call to it in dt_print_ustack() was unguarded, so if the
process died at the wrong instant you could get output looking like this:

                   FUNCTION:NAME
                exit_group:entry
              `<E9><EF>^W`_Exit+0x1d

because dt_Pobjname() returned an undetected error and now you're printing
whatever junk was on the stack at the time.

Fix trivial.  (Everything else in this function is already properly guarded.
This instance has been unguarded since the Solaris days.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>